working gradle+cmake build for android
This solves the broken build regarding onnxruntime for android. The
gradle implementation reference to the AAR apparently downloads
and packages it for android but doesn't make it available to the cmake build
Since we don't know the folder where the doenloaded headers and
lib reside, cmake does its own download. with the advantage that
the cmake should be able to build independently from the gradle build.
That should support native development, testing and so forth later on.